fix #1026, opening deep links (clay menu) in view mode
