+++ lib/preloader/actions.js
@@ -9,6 +9,7 @@ import { PRELOAD_PENDING, PRELOAD_SUCCESS, LOADING_SUCCESS } from './mutationTyp
 import defaultState from './default-state';
 import normalize from '../utils/normalize-component-data';
 import getSites from './sites';
+import parseUrl from './parse-url';
 import { hasAnyBehaviors, convertSchema } from '../core-data/behaviors2input';
 
 /**
@@ -98,41 +99,6 @@ function getPageStatus(state) {
     return 'published';
   } else {
     return 'draft';
-  }
-}
-
-/**
- * Turn the hash string on the location
- * into an object
- *
- * @return {Object}
- */
-function parseUrl() {
-  if (window.location.hash) {
-    const hashProps = window.location.hash.replace('#', '').split('~');
-
-    let propMapping;
-
-    if (hashProps[0] === 'kiln') {
-      // opening the clay menu
-      propMapping = {
-        tab: hashProps[1],
-        sites: hashProps[2],
-        status: hashProps[3],
-        query: hashProps[4] || ''
-      };
-    } else {
-      // opening a form
-      propMapping = {
-        component: hashProps[0],
-        instance: hashProps[1],
-        path: hashProps[2]
-      };
-    }
-
-    return propMapping;
-  } else {
-    return null;
   }
 }
