+++ lib/preloader/mutationTypes.js
@@ -5,5 +5,7 @@ export const LOADING_SUCCESS = 'LOADING_SUCCESS'; // used by view + edit
 export const PRELOAD_SITE = 'PRELOAD_SITE'; // only used by view
 export const PRELOAD_ALL_SITES = 'PRELOAD_ALL_SITES'; // only used by view
 export const PRELOAD_USER = 'PRELOAD_USER'; // only used by view
+export const PRELOAD_URL = 'PRELOAD_URL'; // only used by view
+
 export const META_PRESS = 'META_PRESS'; // only used by edit
 export const META_UNPRESS = 'META_UNPRESS'; // only used by edit
